gnome-session: Split sessions off for use by GNOME Shell package only
GNOME Session by itself is currently used by multiple packages, i.e. by
GNOME and Budgie. However, by itself, GNOME Session is incapable of launching
GNOME Shell.

The support files *should* be part of GNOME Shell, but upstream design is what
it is. As such, we'll make GNOME Shell depend on the subpackage for the session
files, so that we never have unselectable broken sessions.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>