firefox: Update to Firefox 50.0.1
This update addresses CVE-2016-9078. This is not the same as the famed Firefox
0 day update that is currently the topic of discussion, as that exploit is
not executable in this version of Firefox. The Tor browser uses an older
version of the Firefox codebase, so it is affected separately.

CVE-2016-9078 addresses an issue whereby the URL can inherit the wrong
origin after an HTTP redirect, leading to same-origiin domain violations.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>