extra-cmake-modules: Redo stateless patch
Currently it uses /usr/share as the SYSCONFDIR as 99% of files are for xdg so
won't be affected in any way.

This means it's possible to install files straight to /usr/share. The new patch
avoids this by retaining /etc for those files so it's obvious that the source
needs to be patched to use a logical directory.