firefox: Enable e10s out of the box
This change forces Firefox onto the release channel so that it is
considered for e10s. As our Firefox build includes the e10srollout
extension by default, we ensure it's activated first by overriding
the remote autostart property. Testing shows this wasn't always the
case with a fresh profile.

As an aside, we do not disable or alter the extension checking logic,
so if Firefox deems the extensions as non compliant, we'll not force
it on and break anything. With the current beta channel of Firefox
54, multiprocessing is enabled by default, so it stands to reason
that on the next update this pref override is unnecessary, and then
we'll be working with 4 minimum processes.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>