llvm: Update llvm to 4.0.0
This update takes our LLVM package up to version 4.0.0, the latest upstream
version. Additionally this update converts the packaging into the more
modern `ypkg` format.

With this change we do a complete bootstrap of the clang/llvm toolchain,
and build the stage2 from a fresh stage2, ensuring the final resulting
compiler and toolchain (libraries, etc.) are built with a sane clang,
as well as having optimisations applied.

This improves the performance of the libraries and the compiler by orders
of magnitude, and should make Clang a worthy compiler within Solus once
more. As a side effect, anything using the libraries will also see a
significant boost, such as some opensource mesa 3d drivers.

Lastly, the `clang` and `clang-devel` packages have been replaced with
`llvm-clang` and `llvm-clang-devel` respectively.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>