ypkg: Fix pkgconfig private require dependencies in ypkg builds like mutter
Previously we wouldn't account for including the provided pkgconfig
files within the pkgconfig path, so if a .pc depends on another .pc
then we don't include those calculations in.

Effectively, for something like libmutter-0.pc which depends in turn
on mutter-clutter-0.pc, which *then* depends on libxtst.pc, we'd never
end up adding the libxtst.pc dependency to the -devel package. This in
turn leads to having to manually add build dependencies to the package
which is just a pain. This fix ensures we don't have to go the painful
route and automates it.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>