linux-lts: Address CVE-2017-1000364
An issue was discovered in the size of the stack guard page on Linux,
specifically a 4k stack guard page is not sufficiently large and can be
"jumped" over (the stack guard page is bypassed), this affects Linux Kernel
versions 4.11.5 and earlier (the stackguard page was introduced in 2010).

Patch includes initial fix and subsequent fixes.

Signed-off-by: Peter O'Connor <peter@solus-project.com>