openrazer: Update (finally) razer-drivers to 1.1.4
It would appear that razer-drivers is determined to make it virtually
impossible to package for other distributions. As a consequence the
update cadence for razer will have to be reduced in Solus until upstream
stop actively bricking the ability to package it on every update, as
this is proving too hostile a driver to support at this rate.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>