opus: Update to 1.2.1
Opus 1.2.1 is a minor release that fixes a relatively rare issue where the 1.2 encoder would wrongly
assume a signal to be bandlimited to 12 kHz and encode frequencies between 12 and 20 kHz. This only
happens on a few clips, but it is good to update to avoid a potential loss of quality.