nano: Update to 2.8.6
Summary:
GNU nano 2.8.6 "Kekulé" offers a new feature: the ability
to do softwrapping between words -- at whitespace --
instead of always at the edge of the screen.  This can
be activated with -a or --atblanks or 'set atblanks'
together with the softwrap option.  This release further
fixes a handful of rare display glitches, fixes a build
failure on AIX, harmonizes the quoting rules in the rc
files, and renames the option 'cut' to 'cutfromcursor'
(please update your nanorc files before 2020).

[[ https://lists.gnu.org/archive/html/info-gnu/2017-07/msg00007.html | Full Changelog ]]

Test Plan:
  - Edited a file.
  - Gave correct version number.
  - New "-a" (for softwrapping between words) ability is working,

Reviewers: #triage_team

Differential Revision: https://dev.solus-project.com/D678