linux-current: Enable AppArmor `LSM` by default
This takes the Ubuntu 4.12<-4.13 AppArmor backports to provide full
functionality for snapd confining, as well as the introduction of a
fully functioning Linux Security Module within the Solus Linux kernel
build.

This will require some tweaking and may at first introduce some regressions
which we'll look to fix as quickly as possible.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>