bash: Remove all compiler flags overrides from the environment
Controversial point - but I think this is the wrong way for us to "win"
on performance, and tends to cause more issues for users than it really
solves.

Manipulating the compiler for Phoronix tests doesn't actually benchmark
the system packages in Solus (the actual distro) - so is kinda pointless
outside of a bit of frilly PR.

Instead we'll focus on improving the performance of the distro packages
and ensure they provide the best performance, i.e. our Numpy vs some
randomly downloaded Numpy.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>