kernel-glue: Generate mimetype database when there are mime changes
Summary:
Also removed the loop break as no package hits them all, so is inefficient
to check for it each time.

Signed-off-by: Peter O'Connor <peter@solus-project.com>

Test Plan:
Built the plasma ISO, with functional korganizer now mimetypes are
generated as they change on the system. Also invloves rebuilding
`shared-mime-info` with `--disable-update-mimedb` to resolve conflicts
with this package (it no longer ships with a partial db).

Reviewers: ikey

Reviewed By: ikey

Subscribers: folcred

Differential Revision: https://dev.solus-project.com/D1244