jags: Initial Commit of JAGS
Summary: This is the initial commit of JAGS which resolves T4010.

Test Plan: I installed the eopkg and ran the default jags command and looked through the man. I'll be honest its not an indepth test but, I can't find a way to better test JAGS online without knowing more about the program.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Maniphest Tasks: T4010

Differential Revision: https://dev.solus-project.com/D743