bash: Update bash to 4.4.012, move to static build
In an effort to better protect the core of Solus against ABI breakage
during update, we ensure `/bin/sh` is a static binary that is always
available. Using bash arguably isn't the most optimal solution as it
is rather large and slow, but it is the best interim solution to guard
against updates from libs like libreadline.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>