systemd: Fix broken reexec in systemd 235
Some users have the older COMAR script in `kernel-glue` that will force
systemd to `daemon-reexec`. For some of those users this is causing the
entire session to crap out in what can only be describes as the most
unceremionious of fashions.

This patch should fix the broken reexec in systemd.

Upstream [issue](https://github.com/systemd/systemd/issues/7146).

Signed-off-by: Ikey Doherty <ikey@solus-project.com>