nano: Update nano to 2.9.1
Summary:
GNU nano 2.9.1 "Damyatta" fixes a bug where, when the mark is on, ^S would overwrite the file with just the marked region.  This release further clears the "Modified" flag when all edits are undone, adds or updates some magic strings, and does not forget when the cursor was last at line 1, column 1.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Edit this commit message with nano

Reviewers: #triage_team, sunnyflunk

Reviewed By: #triage_team, sunnyflunk

Differential Revision: https://dev.solus-project.com/D1486