ypkg: Update `ypkg` to v26 (T5010)
This new update introduces support for automatic kernel module dependencies,
which will allow us not only to convert the NVIDIA packages to the ypkg
format, but also ensures sane "release==" dependencies.

This means users won't be able to update to the new kernel if they have
the NVIDIA driver isn't ready yet, i.e. they're on unstable and we're doing
kernel rebuilds. Cuz we luv ya.

 - Upstream [changelog](https://github.com/solus-project/ypkg/releases/tag/v26)

Signed-off-by: Ikey Doherty <ikey@solus-project.com>