fontconfig: Downgrade to 2.12.1 because later versions brick Emojis
Yes. Somehow I made this part of my QA criteria today that the ability
to show a full-sized emoji in gtk3-icon-browser was enough to warrant
a downgrade :P

We'll need to reevaluate the fontconfig configuration because it's
certainly showing signs of being busted with newer versions - it stands
to reason this will cause problems elsewhere too (i.e. games)

Signed-off-by: Ikey Doherty <ikey@solus-project.com>