razergenie: Update RazerGenie to 0.5
Summary: This updates RazerGenie to version 0.5. The librazer.so library was renamed to libopenrazer.so (namespace changes will be included in the next release probably) so that's the big abi_symbols change.

Test Plan: RazerGenie still opens. The new "custom editor" opens too. It seems that I have the same bug on Solus as was seen on Ubuntu 16.04, that the QPushButtons don't get colored correctly out of some reason. This needs investigating but is just a visual problem.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.solus-project.com/D1410