veusz: Update veusz to 2.1.1
Summary:
Release notes:
This is a bug fix release. It fixes the error when exporting a plot with dotted/dashed lines to EMF format. The errors when using dataset, tools or import plugins are solved. The splash screen now is hidden after a time period, so it does not obscure dialog boxes. Code to use the deprecated sipconfig if possible when building to get details of the SIP installation has been reintroduced.

Test Plan: Started it followed the tutorial, added a new project, added data to it. Tested some calculations everything worked as expected.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.solus-project.com/D1496