python-iniparse: Initial packaging of python-iniparse
Summary:
python-iniparse is an INI parser for Python which is API compatible with the
standard library's ConfigParser, preserves structure of INI files
(order of sections & options, indentation, comments, and blank lines
are preserved when data is updated), and is more convenient to use.

It is used by the DNF package manager for manipulating its configuration data.

Reviewers: ikey, #triage_team

Reviewed By: ikey, #triage_team

Subscribers: #triage_team, ikey

Differential Revision: https://dev.solus-project.com/D1909