common: Update packages and commit package script
The package script relies on a little manual intervention but can do the job
in about 30s. The latest version allows for the archival of repositories via
adding the repo to ARCHIVEDPKGS. This keeps the git source clean and allows
git history be preserved while scripts can avoid runnning on deprecated
packages (which can give false information about rebuilding a package not in
Solus).

Signed-off-by: Peter O'Connor <peter@solus-project.com>