insomnia: Initial commit of insomnia
Summary:
The inclusion of insomnia fixes T3926

Google Analytics tracking is disabled by default (can still be activated in the settings). I think it is not the expected behavior for a user that a desktop application from the official/trusted repositories is tracking every event in the application.

Depends on D1889

Test Plan:
Start application
Create requests to a REST API

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: darwin67, JoshStrobl

Maniphest Tasks: T3926

Differential Revision: https://dev.solus-project.com/D1785