kodi: Build and include selected distributed-by-kodi binary addons by default.
Summary:
As part of the work on T5112 , make it easy to build and include any
kodi-supplied binary addons requested in the future.

This build also includes and supercedes R1572 kodi-addon-inputstream-adaptive,
which will consequently need to be disabled in the repo.

Kodi 18.0 Leia is slated to exclusively use the cmake build system.

The present version is already configured and built via cmake and
is expected to make updating to 18.0 (currently in RC) as simple as possible.

Test Plan:
Tested on my own HTPC, where I have:

- Scanned for new TV-shows and movies
- Played a h.264 1080p movie and checked that vdpau accel was active
- Used the youtube add-on with the inputstream-adaptive binary-addon to play
  h.264 1080p content with vdpau acceleration enabled (note that settings
  for the inputstream-adaptive plugin need to be manually updated to enable
  a required proxy function for downloading DASH content)

Reviewers: #triage_team, ikey

Reviewed By: #triage_team, ikey

Subscribers: ikey

Differential Revision: https://dev.solus-project.com/D2411