nano: Update nano to 2.9.4
Summary:
GNU nano 2.9.4 "Isabel" allows binding a key to a string (any piece of text and/or commands), permits customizing the color of error messages with 'set errorcolor', colors those error messages by default in bright white on red, makes  at the bottom of the screen scroll just one row when `--smooth` is used, does not fail when redoing a file insertion, and cancels a Shift-selection when any cursor key is pressed without Shift even when the cursor cannot move. Further, it treats tabs fully the same as spaces when doing automatic hard-wrapping, allows syntax names to be unquoted, and removes two deprecated options and six deprecated bindable function names.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: This commit message is written using this version of nano.

Reviewers: #triage_team, joebonrichie

Reviewed By: joebonrichie

Differential Revision: https://dev.solus-project.com/D2472