edk2-ovmf: Initial commit for ovmf, resolves T3214
Summary:
Initial commit for ovmf, resolves T3214.
License in the package.yml field is taken from https://github.com/tianocore/tianocore.github.io/wiki/Legalese.
Might be able pared down because above link inclued all licnenses of the build-system and all packages in the
tar-file. But Not all of them  are shiped in the eopkg file.
Inclueded the license files that are inclueded in the ovmf of the 'common' package.
The secure boot support of ovmf is not inclueded.
Patch is to support more than one thrads to build the basetools.
The build target this uses is inclueded in files. It it a manually modified copy of the template file to build
the ovmf file for our arch. Only jobs is set dynamically. If the next Version changes the target template a new
manually modified version has to be made.

Test Plan:
Use virt-manager to make a new VM. Select 'ovmf bios' in the pre install configuration. Change a variable in the
VMs UEFI and restart see if the change stuck.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake, debug, Timalex, Bunker

Maniphest Tasks: T3214

Differential Revision: https://dev.solus-project.com/D1810