network-manager-applet: Fix double-free in network-manager-applet
This should help with T6298 and get us past the evil crash.

Maniphest Tasks: T6298

This is going to bypass the build system and is being built locally by me
using a clean solbuild image of unstable, so that we can get the emergency
fix out today, sideload onto stable, and bypass the pain that is currently
the libpng rebuilds on unstable.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>