bluez: Extremely evil hack for T5524
The idea here being we'll just exit 0 on the Dell switch which actually
does work, the kernel just tells us it doesn't. We'll use this until we
figure out wtf is going on upstream in the kernel, but at least try to
get the udev portion unbuggered so we're able to load and bind the devices.

Maniphest Tasks: T5224

Signed-off-by: Ikey Doherty <ikey@solus-project.com>