nvidia-glx-driver: Downgrade (temporarily) to 390.46 shortbranch
This goes hand in hand with the kernel reversion, as we're seeing way too
many GPUs being dropped support from the new shortbranch. The only *real*
reason for updating it was to shut up the Feral launcher complaining about
unsupported GPUs that have been supported for a *long* time.

More important that we make sure our users have functioning machines.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>