modelio: Fix T6316
Summary:
Inclusion of this commit fixes T6316.

The desktop files has been modified to rename the menu entry to
"Modelio Open Source" instead of "modelio-open-source" and to enable
pinning of the app.
The new icon has to be included because the former xpm icon file does
not scale properly after addingt the correct WMClass.

Signed-off-by: ahahn94 <ahahn94@outlook.com>

Test Plan:
Installed the resulting package.
Pinning of the application and scaling of the app icon work properly.
Everything seems to work as intended.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T6316

Differential Revision: https://dev.solus-project.com/D2845