python-sqlalchemy: Update python-sqlalchemy to 1.2.17
Summary:
Release 1.2.17 includes a fix to integer handling in the Oracle dialect specific to the 1.2 series, as well as two new event hooks in the ORM to assist with interception of bulk update and delete events.

Full Changelog [here](https://www.sqlalchemy.org/changelog/CHANGES_1_2_17).

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Checked with `pychess`

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Differential Revision: https://dev.getsol.us/D5166