klayout: upgrade klayout to 0.25.8
Summary:
This is a bug-fix release.
Changelog can be found at https://www.klayout.de/development.html

Test Plan: As before, open a few GDS files (kB to GB in size) and browse the layout. Open Help->About and check it ruby/python bindings are present in the `Build options'.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D5941