breeze: Drop our duplicate "default" entry.
To elaborate further, the "default" icon theme was solely added for T3717, effectively overriding the built-in GDK "default" pointer. We now ship a PNG called slick-greeter-pointer.png in the slick-greeter package itself, which we can change to reflect whatever our current or future default theme is, thus eliminating the need for our "default" theme which only introduced confusion.