krename: Switch to git for exiv2 fixes.
Yes, I know this drops locales because we're not using the tarball. I also know that applying a million patches from git on top of the tarball doesn't work because krename seems to use source dating back to seemingly 1942 and doesn't know how to put out a release with all these fixes. Sorry, complain to the krename devs.