syncthing: Update syncthing to 1.3.0
Summary:
This release adds the "databaseTuning" config option to adjust parameters for the database size in use. The default is "auto" which automatically determines appropriate parameters. The other options are "small" (results in the parameters used previously, and used now for small databases) and "large" (results in the new parameters for large databases, regardless of the actual database size).

Bugfixes
- Log spam and high CPU use if #include path is wrong in .stignore
Enhancements
- Rate limit can be very bursty
- Improve free space checking when syncing
- Database performance: improve situation with huge databases
Other issues
- Upgrade server JSON generator
- Allow providing your own ldflags for building

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: Sync a local folder with my NAS. Add & delete file in folder (both side)

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D7299