rustup: Updates rustup to 1.21.1
Summary:
**Changelog**
A panic occurred if a `rustup update` was run with nothing to update and the
download directory was missing. This was harmless but could have confused some
automation jobs.

In release 1.20.x profiles could incorrectly ascribe host-independent components
to the host architecture, resulting in surprising behaviour with `rust-src`.
We have corrected this and added mitigations which should
mean that as of this release, such incorrect ascriptions are supported and also
automatically corrected on toolchain update.

Due to the large number of confusions around the `complete` profile, we have
introduced a warning if you use it. It's really only meant for
developers _of_ Rust, or those exploring particular issues in `nightly`.

There are also a large number of other changes, the highlights of which are below.
Thanks to everyone who helped work on this release. Even if your changes are not
listed below, they are still greatly appreciated.

- Download directory is cleaned up after successful full update.
- Bad `.partial` downloads will be cleaned up for you
- Force installation of toolchain if install is automatic
- Switch to darker colours to improve terminal readability
- Attempt to be less surprising wrt. default-host during installation
- `rustup toolchain list --verbose` now correctly shows the paths
- Fallback environment for non-cargo toolchains updated to match `rustc`
- Made human-readable units slightly more comprehensible
- Improved detection of armhf userland on aarch64 kernels
- Improved error message when rustc is detected on installation

- Added `--profile` support to `rustup toolchain install`
- Added `+toolchain` support to `rustup` itself to match proxy functionality
- Added ability to `rustup component add component-architecture`
- Added clear report when `rustup doc` is run without `rust-docs` available
- Added `keyword:`, `primitive:`, and `macro:` prefix support to `rustup doc FOO`
- Added retry logic so that `rustup` will try and repeat interrupted downloads
- Added `--allow-downgrade` support to `rustup toolchain install`
- Added display of previous version when upgrading channels
- Added support for local non-channel toolchains in rust-toolchain file

Test Plan: Successfully install the stable toolchain by running `rustup toolchain install stable` as a non-root user.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D7871