libmypaint: Remove march=native and mtune=native from cflags
Summary: Passing the `march=native` and `mtune=native` flags are not a good idea on a server. This means that the binaries will be built to be as fast as possible on the server processor and might not work at all or be slower for other people.

Test Plan: - Draw some stuff in mypaint.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D8318