antimicrox: Initial release of antimicroX
Summary: Closes T8638

Test Plan:
Built the package, but noticed I couldn't install the normal and -devel packages together because they had a cyclic dependency. Hmm. Turns out it was because the GUI depends on libantimicro, which is by default grabbed by the -devel subpackage, but -devel depends on the toplevel package, so a cycle happens. That was an easy fix with the `libsplit: no` directive in `package.yml`.

After installing, I opened the app and tested with a couple old profiles I had from the old antimicro and some new ones, everything seems to work fine.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T8638

Differential Revision: https://dev.getsol.us/D8208