gnome-tweaks: Drop Extensions support.
This is [an upstream change currently pending merge](https://gitlab.gnome.org/GNOME/gnome-tweaks/-/merge_requests/50) into GNOME Tweaks. With GNOME 3.36.0, they introduced a new Extensions app that basically just exposes an Extensions list and access to gnome-shell-extensions-prefs, however due to a lot of invasive changes in doing so (which only landed in GNOME 3.36.1) and drop of UUID usage via the gnome-shell-extensions-prefs CLI (along with refactoring show pref logic), it isn't really viable for us to revert that behavior.

Landing this in our GNOME Tweaks is a middle-ground because we actually recognize the value in GNOME Tweaks. It exposes to the user in a more friendly manner many configuration options not just for GNOME Shell, but also for Budgie 10 series (depending on the location / functionality). The Extensions app literally just lists extensions and opens up the same old preferences dialog that Tweaks was calling, its value is limited. So we're not deprecating Tweaks as a whole.

tldr; GNOME unsurprisingly regressed behavior in a new release of something.