imlib2: Update imlib2 to 1.6.1
Summary:
The newer versions adds primarily adds support for the WEBP format and lots of bugfixes.

You can see the full ChangeLog in the tarball of this version (couldn't find a link to it).

Signed-off-by: Martin Reboredo <gc1000ll@gmail.com>

Test Plan: Saved an image of WEBP format in ANSI and SVG formats with libcaca (`img2txt`) and displayed the outputs in `less -r` and `eog`.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D8731