ckb: Fixes #T8734 Can't apply udev rules when ckb-next-daemon is running
Summary:
usysconf fails reloading udev rules (udevadm trigger) when ckb-next-daemon is running.
As written in the bug report this is a kernel issue, and the ckb-next developers
decided to work around it since it takes time for kernel maintainers to fix and for
the kernel to appear downstream on the distributions.

I decided to apply the relevant patches to 0.4.2 to fix this issue, this patches should not
be necessary any longer once 0.4.3 is released.

Test Plan:
* Update the package (with a corsair peripheral)
* restart ckb-next-daemon
* run udevadm trigger
* No more error messages are displayed, thus, usysconf won't complain any longer.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D8855