guitarix: Update guitarix to 0.39.0
Summary: No news provided by upstream. I was going to push this as part of the gnome stack upgrade but it got lost in my package folder for some reason. Here it is, a bit late, but better late than never.

Test Plan: - Use the standalone application and lv2 plugins without issues.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D8852