gsettings-desktop-schemas: Default to touchpad being enabled while typing.
The default of the touchpad being disabled makes navigating around Budgie and GNOME (especially in the Activities view) while using keyboard shortcuts unnecessarily difficult and there is no obvious method of disabling it via GNOME Control Center (requiring the use of Tweak Tools and the knowledge that it is an option in the first place).

Also makes it difficult to play games with the touchpad. Not sure why you'd want to subject yourself to that but hey, it's your gaming experience. Resolves T8948.