chocolate-doom: Update chocolate-doom to 3.0.1
Summary:
**Changelog:**
This is a point release that fixes a security vulnerability (CVE-2020-14983). An unchecked field in the Chocolate Doom server logic could allow a malicious attacker to trigger arbitrary code execution against Chocolate Doom servers.

**Packaging changes:**
1. Enable support for PNG screenshots with libpng - by default Chocolate Doom uses ancient PCX format (as well as the original game in 1993) that isn't even properly supported by some modern image viewers.
2. Enable libsamplerate for high-quality sample rate conversion - original Doom sounds were sampled in unsigned 8 bits, mono channel and frequency of 11 KHz or 22 KHz; this functionality allows the game to resample these sounds to signed 16 bits, stereo channels and system-default frequency.

Test Plan: Completed MAP01 and MAP02 from Doom 2. Configured the game to use PNG format for screenshots, made a screenshot and viewed it in Gwenview and EOG.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9563