bottom: Initial inclusion of Bottom
Summary:
This package doesn't replace ytop as it's really more of an alternative than a direct replacement. Given the future of ytop in the repo, the end result is pretty much the same, anyways.

Closes T9265

Test Plan: Run btm and make sure that all of the system information displays correctly and appears to be accurate.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T9265

Differential Revision: https://dev.getsol.us/D9567