mojave-gtk-theme: sync with git
Summary:
**Changelog:**

- update gnome-shell popover message style
- Fixed login username issue
- Fixed gdm issues
- Fixed gnome-shell issues

**Packager notes:**
This theme is actively developed but there isn't a new release.
Latest release is version 20200324, but it had issues with the text color for file names on the desktop.
Sync with git since these issues are fixed.

Signed-off-by: Algent Albrahimi <algent@protonmail.com>

Test Plan: Install locally in Budgie and Gnome DE and switch system theme to all Mojave variants.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9867