mpv: backport patches for better compatibility with libass 0.15.0
Summary:
The first patch corrects the rendering of the OSC - it relies on `ScaledBorderAndShadow` being set to 1, and as in new libass the default has been set to 0, this property needs to be set explicitly.
The second patch is purely cosmetic - it just replaces deprecated funtion calls with modern ones; the functionality shouldn't be affected, but we won't get deprecation warnings during build.
Depends on D9887.

Test Plan: Watched video with complex ASS/SSA subs in mpv, checked the correctness of the subtitles and OSC.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D9888