yt-dlp: Switch to yt-dlc fork.
Due to youtube-dl no longer being made available as a result of a DMCA request by the RIAA to GitHub, we are now moving to the more actively maintained (and still available) yt-dlc fork.

This fork addresses issues raised in regards to copywritten content being referenced in youtube-dl's test suite. yt-dlc does not have these tests and thus such issue should be resolved.

As a reminder, Solus does not condone the use of yt-dlc for copyright violation and circumvention. Its place in our repository is on the basis of ensuring the ability for users to archive fair use and Creative Commons content.