budgie-desktop: Fix workspace switching for default window in an IconTasklist.
This commit fixes a bug that was the result of using the DesktopHelper's get_stacked_for_classgroup function. This function would only return a list of windows on the current workspace, rather than on all workspaces.

Instead, we will now using the tracked apps in our IconPopover, since that also leverages Abomination's filtering, and create a list from that. If we only have one window, or more than one with the show-all-windows-on-click set to be disabled, we'll proceed to check if the window is on the workspace. If the window is not on the active workspace, we'll activate that workspace then activate and unminimize the window. It's likely that if you're wanting to switch to a workspace to get to a window to begin with, chances are you probably want to actually see it. If the window is on the active workspace, we'll toggle its state.