budgie-desktop-view: Ensure we create the Trash files directory if it doesn't exist.
This can happen on a fresh install or an install that simply has never had files deleted via the file manager before. We'll provide a warning notifying that it doesn't exist, attempt to create it, and provide a warning if we fail.

Added a null checks for the trash_item so we're not trying to perform operations if the item doesn't exist.