budgie-desktop-view: Update to 0.99.3.
**Changelog:**

Implement right-click menu for FileItems.

I've implemented a dedicated right-click menu for FileItems and introduced a new gsettings key called "terminal" since there is no mechanism to actually "detect" what the default Terminal is. By default this is set to gnome-terminal, however an end user can set it to basically anything, though we only support the following out-of-the-box:

- alacritty
- gnome-terminal
- konsole
- mate-terminal
- terminator
- tilix

I documented with fairly extensive detail the flags supported by each Terminal application that we "support" and provide a warning (if you're running the desktop view process directly or checking logs) when you're trying to use a Terminal that isn't supported. Open to adding more, just need more testing.

This "Open in Terminal" option also does not hold the process open. It's my judgement that such preference such be left up to both the application and the user, for example in GNOME Terminal there is an option under Preferences -> whatever your default Profile is -> "When command exits" that decides what to do.

This implementation also comes with the caveat that GNOME Terminal is special. It's no mystery, but in this case it won't successfully set the working directory if you have a custom command set to run instead of your shell, and if you have "Preserve working directory" set to anything besides "Always". Bare that in mind.

I've eliminated the child_activated handler on the FlowBox children. MountItem has its own left-click handling as a button_release handler, with no actions on right click. Special directories use the FileItem right-click menu. For FileItems that are applications, we prevent the FileItem menu from showing. This may be changed in the future.