fontforge: Update fontforge to 20201107
Summary:
This release falls on the 20th anniversary of the first release of FontForge back in 2000. It brings a wide range of minor tweaks and bug fixes for the user interface and file format handlers and a special splash screen to commemorate the big day.

Depends on D10116 D10117

Signed-off-by: Algent Albrahimi <algent@protonmail.com>

Test Plan: Draw a new font and modify a few fonts.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D10119