gnome-control-center: Bump for 3.36.x release inclusion.
Reverting back from 3.38.x because at some point in GNOME Control Center 3.38 development cycle, they broke most of the display configuration and haven't been responsive to bug reports file by Solus users on the matter.

Internally we will be going through the internal releases to determine when it broke and bisect the specific commit to resolve it.