wine: Update wine to 6.0
Summary:
This release represents a year of development effort and over 8,300 individual changes. It contains a large number of improvements that are listed in the [release notes](https://www.winehq.org/announce/6.0) below. The areas of major changes are:
- Core modules in PE format.
- Vulkan backend for WineD3D.
- DirectShow and Media Foundation support.
- Text console redesign.

Signed-off-by: Pierre-Yves <pyu@riseup.net>

Test Plan: wine64 is working fine, but I need some help to test & validate the 32-bit version.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl, YakoYakoYokuYoku

Differential Revision: https://dev.getsol.us/D10369