budgie-desktop-branding: Introduce fortitude-plus sub-package for Solus 4.2
The purpose of this new Solus 4.2 sub-package is to introduce changes that bring it in line with our defaults for GNOME and Plasma around click policy (going from double to single click) and enabling Desktop Icons by default. This is done in a manner that will not affect existing Solus 4 / 4.1 users.