rizin: Switch from radare2 to rizin
Summary:
Cutter has switched from r2 to Rizin as its backend, so we have to make this switch to continue supporting it.
Rizin also seems to be more packaging-friendly project in the long term:
- Meson is now used as a default build system instead of custom ACR system of r2.
- Rizin aims to provide a C API for external projects instead of relying on string commands.

Other interesting changes include proper project saving and loading (now it saves and loads the state of the reversing environment instead of simply replaying the commands) and usage of tree-sitter-based command parser by default.

Changelogs:
- [0.1.0](https://github.com/rizinorg/rizin/releases/tag/v0.1.0)
- [0.1.1](https://github.com/rizinorg/rizin/releases/tag/v0.1.1)
- [0.1.2](https://github.com/rizinorg/rizin/releases/tag/v0.1.2)

Depends on D10789.

Test Plan:
1. Disassembled /bin/ls with Rizin.
2. Rebuilt and tested Cutter 2.0.0 against it.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D10787