syncthing-gtk: Switch to Python 3 Port
Summary:
The original author decided not to update to Python3 (and has effectively abandoned the project). This port is de facto the port that the community appears to have rallied behind and that has been used to reintroduce the package into Debian.

Part of T8207

Test Plan: Shared a folder. Verified that folder synced. Checked that the appindicator and notifications were working. Checking that no UI elements appear to be broken.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl, algent

Differential Revision: https://dev.getsol.us/D11192