protonvpn-cli: Update protonvpn-cli to official cli version 3.8.0
Summary:
Before we've been using the community version, which contains less feature than the official cli (e.g. kill switch). From the 10th release, we will be using the official cli.

Unlike other distributions, the Proton python client and the ProtonVPN network manager library are installed in the "setup" section, instead of as separate packages in the repo. If in the
future, someone wants to package the GUI, use the "setup" section and the same dependency (excluding pythondialog).

Changes:
1. Changed multiple dependencies, because this is to some extent a new package.
2. The command changed from `protonvpn` to `protonvpn-cli`. Basic sub-commands, such as connecting and disconnecting, are still the same.

Test Plan:
Tested on a Solus VM under the unstable repo. Tested login, ran nspeedtests, and connected to multiple locations.

Signed-off-by: Gavin Zhao <gavinzhaojw@protonmail.com>

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D11670