xorg-server: Drop cvt files
These are to be replaced with a standalone library libxcvt. The upcoming
xorg-server will hard-depend on this library.

As libxcvt will be a dependency of xorg-server and gets installed first
the installation will abort due to file conflicts. Do some packaging
gymnastics to avoid this problem ahead of the xorg-server update.