nextcloud-client: Update nextcloud-client to 3.4.1
Summary:
This release fixes a major regression introduced in 3.4.0 where files with a unix time of 0 would be created, would spread to other clients, and trigger resyncs on those clients.
More details can be found [here](https://help.nextcloud.com/t/desktop-client-3-4-0-destroys-local-time-stamp-and-keeps-uploading-data-to-server/128512).

Release notes:

- [v3.4.1](https://github.com/nextcloud/desktop/releases/tag/v3.4.1)
- [v3.4.0-rc1](https://github.com/nextcloud/desktop/releases/tag/v3.3.6), [v3.4.0-rc2](https://github.com/nextcloud/desktop/releases/tag/v3.4.0-rc2) (v3.4.0 has been removed).
- [v3.3.6](https://github.com/nextcloud/desktop/releases/tag/v3.3.6)

Test Plan:
Start the client in an environment affected by the aforementioned issue and fixed on the server side.
Verify that a) local timestamps are fixed and b) no incorrect files are uploaded to the server.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Differential Revision: https://dev.getsol.us/D12521