leiningen: The inclusion of leiningen fixes T10039
Summary:
this will package leiningen, it makes some modifications to download
the required files to a location other than home in the build, it also
change the path to a system path instead of looking for the jar in the
users home folder.

Test Plan:
Ran make to build the package and run eopkg install to install from the resulting file
Ran "which lein" and "lein help" in a new terminal to check for version and output.

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: DataDrake

Differential Revision: https://dev.getsol.us/D12459