paperwork-backend: Initial commit version 2.1.0
Summary:
Depends on D12701
Depends on D12702

This version removes python-psutil as direct dependency as it comes from
openpaperwork-core.

It depends now on scikit-learn.

The tests are not working, because some files are missing and the tests
are using some mechanism that is blocked by ypkg, I suppose...

Test Plan:
Build paperwork-gtk/paperwork and it works.
(Because of libhandy it is yet to be done.)

Reviewers: #triage_team, DataDrake

Reviewed By: #triage_team, DataDrake

Subscribers: algent, DataDrake

Differential Revision: https://dev.getsol.us/D12703