akonadi: Add the MariaDB backend package
Summary: Add the MariaDB backend package. This is an effort to make SQLite3 the default database backend. For now, we're setting SQLite3 as the default backend, while adding the MariaDB support as a dependency. This is to avoid breaking current installations. See T9922.

Test Plan: Both backends work when set in `~/.config/akonadi/akonadiserverrc`.

Reviewers: #triage_team, Girtablulu

Reviewed By: #triage_team, Girtablulu

Subscribers: Girtablulu

Differential Revision: https://dev.getsol.us/D13171