gdk-pixbuf: Remove pre-compiled loaders.cache file
With the addition of webp-pixbuf-loader into the repo, a trigger
was added to usysconf to dynamically update the loaders.cache file as
new loaders are added.

However, for emul32 as there are no additional packages requiring
a dynamically updated cache file, we can keep relying on the pre-compiled
loaders.cache file.