pipewire: move FFmpeg plugin to subpackage
Summary: I'm proposing to move this plugin to subpackage to mitigate the dependency cycle that I stumbled into in D13904. Only one SPA plugin in `pipewire` package depends on `ffmpeg`, and this plugin isn't necessary for general PipeWire operations, so it can be safely moved to another package and dependency cycle `sdl2`->`pipewire`->`ffmpeg`->`sdl2` can be avoided.

Test Plan: Rebuilt `sdl2` with static dependency on `pipewire` and checked that `sdl2`, `ffmpeg`, `pipewire` and `pipewire-spa-ffmpeg` packages are co-installable and there is no cycle dependency between them.

Reviewers: #triage_team, Staudey

Reviewed By: #triage_team, Staudey

Subscribers: Staudey

Differential Revision: https://dev.getsol.us/D14113