libplacebo: Update libplacebo to 6.292.1
This minor release includes several important bug fixes to downscaler kernels, a new and improved default perceptual gamut mapping mode, and an overhaul of the `plplay` demo program's core timing loop. It also drops the requirement for VK_KHR_synchronization2, which was unnecessarily and prohibitively strict in practice.
Note that because this minor release does not change the public API, several new options related to the above changes are not yet exposed as part of this release.

Full changelog [here](https://code.videolan.org/videolan/libplacebo/-/tags/v6.292.1).