protonmail-bridge: Update protonmail-bridge to v3.3.2
Summary:
The build system changed a lot compared to the last update I did and I had to patch a number of things to make it work. Since the last release one of the binaries was split into `bridge` and `bridge-gui` which I included as `proton-bridge` and `proton-bridge-gui`. The launcher application remains as `protonmail-bridge`.

The new build uses `vcpkg` and I included port overlays for the dependencies to fall back on the system libaries. Because `vcpkg` is expected as a submodule I had to switch the build to the git tag.

**Summarized changelog:**

- 3.3.0
  - Added new notifications to help users to configure and troubleshoot their email clients
- 3.2.0
  - Enhanced the integration with the operating system by replacing status windows with native tray icon context menu
- 3.1.2
  - Support for Protons Scheduled send feature
  - Synchronization indicator in the GUI
- 3.0.20
  - Fixed a bug that could cause email states to come out of sync with the web application
- 3.0.17
  - Significant rewrite of bridge and update to Qt6
  - Improved sending and synchronization performance
  - Added subfolder support

Full changelog: https://proton.me/download/bridge/stable_releases.html

Test Plan: Synced and sent mail through Thunderbird.

Reviewers: #triage_team, EbonJaeger

Reviewed By: EbonJaeger

Spies: EbonJaeger, ReillyBrogan

Differential Revision: https://dev.getsol.us/D14293