nvidia-glx-driver: Fill out modaliases, include 32bit driver appstream data.
For the driver to actually show up in gnome-software then the provided modaliases,
must match to the hardware installed. Use a script from rpmfusion to do this.

Additionally, also create a 32bit addon for the 32bit subpackage that extends,
both -lts and -current variants. This will show as an addon in gnome-software.

Finally add a hard requires on the relevant kernel version for LTS and current
variants for supported software centers.