font-terminus-console: Initial inclusion
**Summary**
Add the console version of the terminus monospaced font.

The intent is for this to be used in the bootloader, the kernel and by
default in the linux console (ter-vXXn), which will make things look
particularly sleek.

**Test Plan**
Install font-terminus-console locally and use setfont to switch to
ter-v32n and observe that it looks nice on a 32" 4K monitor.

Signed-off-by: Rune Morling <ermo@serpentos.com>