grub2: Add terminus font .pf2 files
**Summary**
This commit adds terminus BDF format fonts to the grub package recipe,
and auto-generates grub native .pf2 fonts as part of the package build
process.

Doing it like this serves to show both how to generate grub fonts and to
test that grub-mkfont actually works on each build of the grub2 package.

In addition, this revision to grub2 is part of the drive to ensure that
we can use the same font face from boot menu to linux console, cf. #826

Signed-off-by: Rune Morling <ermo@serpentos.com>