lightdm: Check for membership in the users group for autologin
Formerly we were checking to see if a user was in the autologin group to see if they were allowed to login. Solus doesn't have the autologin group, and while we could create it it doesn't appear that accountsservice or Plasma have functionality to automatically add a user to that group when autologin is enabled.

Since autologin already requires root writing permissions (due to writing to /etc/lightdm/lightdm.conf) checking if a user is in autologin is redundant for purposes of security. Let's just change it to a check to see if the user is in the users group since those are the only users that should be doing interactive login anyway.

Part of getsolus/packages#876

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>