ci/package_checks: fix SPDX license check for WITH/AND/OR
SPDX license identifiers may be a 'simple' or 'compound' expression,
meaning that the following formats are also allowed:

- A document reference instead of license identifier.
- `+` after the license identifier.
- License identifier followed `WITH` and an exception identifier.
- The above split by `AND` or `OR`.
- The above contained between `()`.

All but the first case should now be valid according to the package checks.
The first case is still invalid, as we want license identifiers and not document references in our packages.

See: https://spdx.github.io/spdx-spec/v2.3/SPDX-license-expressions/

Resolves #956