grub2: Update default config to set font to ter-32b
**Summary**
Increase default font size to 32px (terminus bold) to help with
legibility on HiDPI displays.

Additionally, set GRUB_GFXMODE 1366x768, with a 1024x768 fallback, and
auto for when VM gfx adapters are genuinely clueless. This appears to
be a decent setting for everything from VMs and older laptops to modern
high resolution displays with the added bonus that the font itself is
cleaner looking than the one upstream ships by default.

Part of #826

**Test Plan**
- Update grub2
- run `sudo mv /etc/default/grub.newconfig /etc/default/grub`
- run `sudo clr-update-wrapper`
- reboot and check that new font is displayed in grub bootloader menu

Signed-off-by: Rune Morling <ermo@serpentos.com>