pam: Set more tolerable failed login defaults
**Summary**
This commit tweaks the defaults before an account is locked from:

3 unsuccessful tries within 15 minutes => 10 minute account lockout

To:

10 unsuccesful tries within 15 minutes => 10 minute account lockout

The goal is to make fat-finger typos less punishing and frustrating. The
hope is that allowing more consecutive tries before locking the user out
will not punish legimate mistakes, but will still punish an attacker
trying to brute-force the user's password.

Note that the NIST recommends allowing at least 10 tries here:

https://pages.nist.gov/800-63-3/sp800-63b.html#sec10

Signed-off-by: Rune Morling <ermo@serpentos.com>