grub2: Improve default resolution settings
**Summary**
On BIOS-based virt-manager VMs, the grub2 graphics mode defaults to VGA
640x480 if no configuration changes are made to the `GRUB_GFXMODE` stanza
in `/etc/default/grub`.

However, it turns out that the BIOS on the emulated VGA card supports
the 1280x800 mode, which makes the default ter-32b font we use much more
usable compared to when it is used in 640x480.

At the same time, 1280x800 is a resolution that is so unusual, that
almost no physical displays and almost no real world VGA BIOSes support
it.

And if -- against expectations -- a physical GPU+Display combination
turns out to support 1280x800, the resolution is large enough that the
ter-32b font will still yield a 80x25 display and look nice in doing so.

Signed-off-by: Rune Morling <ermo@serpentos.com>