krita: Fix gmic crash
The gmic plugin was building against a mixture of system gmic headers and included gmic src. This caused crashes due to the included gmic src being somewhat older than the system version. Fix this by deleting the included gmic src so that it can't be used (this also requires a fix to system gmic which has already been applied).

Closes getsolus/packages#1477

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>