restic: Update to 0.16.4
**Changelog:**

This release works around and improves detection of a bug in the compression library used by restic.
The resulting issue only happens when using restic 0.16.3 and the max compression level (the default auto and off compression levels are not affected),
and when the source files being backed up have specific data in them to trigger the bug.

If you use max compression, you can use restic check --read-data to make sure you're not affected.

Signed-off-by: Troy Harvey <harveydevel@gmail.com>