sddm: Build qt5-greeter and update inputmethod
inputmethod=qtvirtualkeyboard causes a large on-screen keyboard to appear for non-breeze themes. Disable that until it's resolved upstream.

Also, build a qt5 greeter for SOME support for qt5 themes. Note that anything that depends on Plasma components will fail with an error, but a few themes that do not are likely to work.

Resolves getsolus/packages#2040

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>