ypkg: Control which binaries to use via symlinks
**Summary**
This commit renames the pure python executables to name.py and then names
the nuitka-compiled binaries name.bin.

In turn, it symlinks name to name.bin as a matter of packaging policy,
which in turns enables us to trivially test and/or switch to/from the
pure python implementations by default via a simple release bump.

This has the benefit of always giving people the ability to inspect the
source code of the current version of ypkg residing on their systems, and
enables Solus to trivially do performance tests of the pure python
versions vs. the nuitka compiled versions for any given package release.

Signed-off-by: Rune Morling <ermo@serpentos.com>