llvm: Make .def files architecture-dependent
These files contain the various settings that llvm was built with. Previously the 64-bit versions of these files would override the 32-bit ones, which broke the mesalib 24.1-rc1 build as the 32-bit build of that would try to link against symbols that were non-existent in the 32-bit version of libLLVM-16.so.

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>