bash: Build new PATH the smart way
**Summary**
The goal with this PR is to ensure that any existing user-set PATH is
respected to the widest extent possible. Any elements of the default PATH
which have not already been added, will be added to the new PATH in (what
I hope to be) a sane way.

In the case of no user-set PATH, prepare a sane default PATH.

I have left the debug statements in for now to provide visibility to
the various test cases. Naturally, these debug statements will need to
disappear before the commit is merged to main.

Signed-off-by: Rune Morling <ermo@serpentos.com>