pisi: Order baselayout first, don't reorder eopkg
**Summary**
This should fix issues with eopkg being uninstallable because glibc was
not installed first.

Note that this release adds a rundep on the eopkg package to ensure eopkg
always gets installed before pisi. For the epoch bump, ownership of
certain files in the pisi package will be transferred to eopkg, which
makes the rundep even more pertinent.

Signed-off-by: Rune Morling <ermo@serpentos.com>