libuiohook: Update to 1.2.2
**Summary**
No changelog provided.

Part of https://github.com/getsolus/packages/issues/3554

**Packager note**

This was added for an OBS plugin, but I don't see it in the repository, nor does anything pull it in as a dependency, so I can't tell if this is used. I also don't see any matching deprecated packages.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>