weston: Split package into parts
**Summary**
- Splits Weston between `weston`, `weston-libs`, `weston-session`, and `weston-demos`

Doing this is useful in the cases where Weston is being used as just a compositor, such as SDDM. In that case, it's not really useful or helpful to have the demo programs and session files installed. If someone decided to make a new compositor based on Weston, they could then also do that without having to pull in all of Weston.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>