solbuild: call ypkg-install-deps w/ --eopkg-cmd arg
By default, this will be eopkg.bin, which in turn means that the solbuild
images should be built w/eopkg.bin, so the FilesDB doesn't continuously
get rebuilt on solbuild build invocations due to the new eopkg FilesDB
auto-upgrade code.

The end result is that we will no longer be using eopkg.py2 in our build
pipeline. Yay.

Signed-off-by: Rune Morling <ermo@serpentos.com>