lact: Support Solus kernel argument
It looks like lact was directly checking to see if the power features mask was set instead of checking to see if overclocking features were available. Patch it so that it detects our new amdgpu kernel parameter correctly and also patch out that pesky initramfs generation step since it doesn't work on Solus anyway.

For getsolus/packages#4032

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>