calamares: Explicitly set /etc/vconsole.conf font
**Summary**
As part of the CBM changes in [1], CBM will now parse the system
`/etc/vconsole.conf` file and explicitly set the corresponding

`rd.vconsole.font=(...) rd.vconsole.keymap=(...)`

options on the kernel commandline, such that they are used in rescue
initrd shells.

[1]: https://github.com/getsolus/clr-boot-manager/pull/7

To make it clearer and more discoverable to the user what these variables
are currently set to in the root fs, the present change ensures that the
Solus default font selection of the terminus 32px bold console font with
unicode support (ter-v32b), will now also be explicitly listed on the
kernel command line for troubleshooting/rescue purposes via CBM.

This is primarily an ergonomics change meant to help Solus users
understand how to influence these settings in the face of error modes
that prevent them from logging in via their normal graphical session.

The consensus seemed to be that this PR ought to be followed up by a
Documentation PR on the help center, expanding on the options between
which the user can choose in terms of font sizes, font styles, and
keyboard mappings, should they need to change the settings for whatever
reason.

Signed-off-by: Rune Morling <ermo@serpentos.com>