golang: Update to v1.24.3
**Summary**

go1.24.3 (released 2025-05-06) includes security fixes to the `os` package, as well as bug fixes to the runtime, the compiler, the linker, the `go` command, and the `crypto/tls` and `os` packages. See the [Go 1.24.3](https://github.com/golang/go/issues?q=milestone%3AGo1.24.3+label%3ACherryPickApproved) milestone on the issue tracker for details.