common/Scripts: Add a script to validate AppStream
Adds a script to validate an AppStream metainfo XML file. It uses a
bunch of overrides sourced from Flathub. I decided to use these to try
to maintain the best compatibility; the idea is if a metainfo file works
for Flathub, it should work for us.

Some exceptions are commented out as they rely on patches to appstream
(I think?), and I haven't looked into locating those yet. We can decide
to pursue that at a later time and uncomment them, or just drop them.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>