pandoc-crossref: Add runtime dependency on pandoc
**Summary**

When updating `pandoc` or `pandoc-crossref` we must ensure that
`pandoc-crossref` is built against the same version of `pandoc`
despite not directly using it as build dependency.
Otherwise, if versions are mismatched `pandoc-crossref` will report a warning:
```
WARNING: pandoc-crossref was compiled with pandoc 3.7.0.2 but is being run through 3.6.3.
This is not supported. Strange things may (and likely will) happen silently.
```

So since `pandoc-crossref` is meant to be used only with fixed version
of `pandoc` from repos, it makes sense to add runtime dependency between
them.