hashcat: Update to v7.1.2
**Summary**
This hotfix restores backward compatibility in machine-readable status view mode, which was broken in v7.1.0 and affected third-party overlays such as the Hashtopolis agent. It also fixes issues in the Argon2 libraries when used in multi-hash modes. If you are using hashcat v7.1.0+ with LUKS2 or KeePass KDBX4, updating is strongly recommended.

The release also introduces a new assimilation bridge plugin, the Rust bridge. Just like the Python bridge plugin, which lets you write hash-mode extensions in Python instead of C, you can now do the same in Rust. This is a valuable community contribution that will be covered in more detail in the next release notes. The feature is already fully functional and ready for early adoption, with initial documentation available in hashcat-rust-plugin-requirements.md.

For the full list of changes, please see docs/changes.txt.
Resolves: https://github.com/getsolus/packages/issues/6650