librsvg: Disable autodeps
**Summary**
This is so we can avoid a circular dependency issue between
librsvg/gdk-pixbuf/glycin.

We keep deps as before, just excluding the dependency on gdk-pixbuf.

As gdk-pixbuf depends on glycin and glycin deps on librsvg, this
should nicely resolve the issue as glycin will pull the librsvg
dependency for gdk-pixbuf regardless like that.